Moved operation on register from cmd
