+++ CMakeLists.txt
@@ -16,8 +16,10 @@ ADD_SUBDIRECTORY(src/memory)
 ADD_SUBDIRECTORY(src/elf)
 ADD_SUBDIRECTORY(src/command)
 ADD_SUBDIRECTORY(src/cli)
+ADD_SUBDIRECTORY(src/misc)
 
 TARGET_LINK_LIBRARIES(${MYGDB} PUBLIC ${MYGDB}_memory)
 TARGET_LINK_LIBRARIES(${MYGDB} PUBLIC ${MYGDB}_elf)
 TARGET_LINK_LIBRARIES(${MYGDB} PUBLIC ${MYGDB}_command)
 TARGET_LINK_LIBRARIES(${MYGDB} PUBLIC ${MYGDB}_cli)
+TARGET_LINK_LIBRARIES(${MYGDB} PUBLIC ${MYGDB}_misc)
